change element collection->entity relational mapping
