+++ src/main/java/com/seneca/senecaforum/client/controller/SidebarController.java
@@ -1,6 +1,6 @@
 package com.seneca.senecaforum.client.controller;
 
-import com.seneca.senecaforum.domain.SidebarMenu;
+import com.seneca.senecaforum.domain.ui.SidebarMenu;
 import com.seneca.senecaforum.repository.SidebarRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
