+++ src/main/java/com/seneca/senecaforum/domain/ui/Topic.java
@@ -1,0 +1,31 @@
+package com.seneca.senecaforum.domain.ui;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.OrderBy;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "topic")
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Topic {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "topic_id")
+    private int id;
+
+    @Column(name = "topic_name")
+    private String topicName;
+
+    @Column(name = "views")
+    @OrderBy(clause = "desc")
+    private int views;
+
+}
