+++ src/main/java/com/seneca/senecaforum/repository/SidebarRepository.java
@@ -1,6 +1,6 @@
 package com.seneca.senecaforum.repository;
 
-import com.seneca.senecaforum.domain.SidebarMenu;
+import com.seneca.senecaforum.domain.ui.SidebarMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SidebarRepository extends JpaRepository<SidebarMenu,Integer> {
