+++ src/main/resources/application.yml
@@ -4,10 +4,9 @@ spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://localhost:3306/seneca_forum
-#    username:
-#    password:
+    username:
+    password:
   jpa:
-    open-in-view: false
     show-sql: true
     database-platform: org.hibernate.dialect.MySQL5Dialect
     hibernate:
@@ -26,3 +25,5 @@ spring:
 #  resources:
 #    static-locations: classpath:/videos/
 #secretKey: MjlhMTU0MzFiNmUxYzY0MDlkMGYwM2ZhMTE5NGYyNWI1YTIxOTgxMjE5MGYyYzU1NGU2OGM4ZjgxNjRlOWIwN2IxODYxYWY2OGFjNmQ0ZDk2YjY2MDdmYmYzMWZkMGUzNjQzOWU1M2YxODQ2NmM0MTIzNGIwZDE3NWVlZWIzYTk=
+
+
